requirements.txt
Flask==2.3.3
flask-cors==4.0.0
boto3==1.34.59
python-dotenv==1.0.1

deepface==0.0.79
numpy==1.23.5
scikit-learn==1.3.2
opencv-python==4.8.1.78
Pillow==10.3.0

tensorflow==2.12.0
keras==2.12.0

Werkzeug==2.3.7
